Building ProductReviewActivity

+++ app/src/main/java/com/pekwerike/madeinlagos/repository/MainRepositoryAPI.kt
@@ -11,7 +11,7 @@ interface MainRepositoryAPI {
     val allProductsWithReviewsAsLiveData: LiveData<List<Product>>
     suspend fun refreshProductList(): NetworkResult
     suspend fun getProductReviewsByProductId(productId: String): NetworkResult
-    suspend fun postProductReview(productId: String, userRating: Float, userReviewText: String): NetworkResult
+    suspend fun postProductReview(productId: String, userRating: Int, userReviewText: String): NetworkResult
     suspend fun getProductById(productId: String): Product
     suspend fun fetchAllProducts(): ProductsAndNetworkState
 }
